+++ include/mruby_integration/core/window.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "mruby.h"
-
-void
-append_core_window(mrb_state*);