+++ src/mruby_integration/core.cpp
@@ -6,7 +6,6 @@
 #include "mruby_integration/core/input.hpp"
 #include "mruby_integration/core/misc.hpp"
 #include "mruby_integration/core/timing.hpp"
-#include "mruby_integration/core/window.hpp"
 
 #include "ruby/core/buildkite_analytics.hpp"
 
@@ -19,7 +18,6 @@ append_core(mrb_state* mrb)
   append_core_input(mrb);
   append_core_misc(mrb);
   append_core_timing(mrb);
-  append_core_window(mrb);
 
   load_ruby_core_buildkite_analytics(mrb);
 }